chore: move sortBy, uniqBy to utils
