+++ lib/v2/showstart/service.js
@@ -1,5 +1,5 @@
 const { HOST } = require('./const');
-const { getAccessToken, post } = require('./utils');
+const { getAccessToken, post, sortBy, uniqBy } = require('./utils');
 
 async function fetchActivityList(
     params = {
@@ -68,30 +68,6 @@ async function fetchStyleList() {
         link: `${HOST}/event/list?showStyle=${item.key}`,
         description: `showStyle: ${item.key}`,
     }));
-}
-
-// sort by key
-function sortBy(items, key) {
-    return items.sort((a, b) => {
-        if (a[key] < b[key]) {
-            return -1;
-        }
-        if (a[key] > b[key]) {
-            return 1;
-        }
-        return 0;
-    });
-}
-
-function uniqBy(items, key) {
-    const set = new Set();
-    return items.filter((item) => {
-        if (set.has(item[key])) {
-            return false;
-        }
-        set.add(item[key]);
-        return true;
-    });
 }
 
 async function fetchDictionary(cityCode, showStyle) {
