+++ lib/v2/showstart/utils.js
@@ -60,8 +60,33 @@ const post = async (requestPath, accessToken = md5(Date.now().toString()), paylo
     return response;
 };
 
+function sortBy(items, key) {
+    return items.sort((a, b) => {
+        if (a[key] < b[key]) {
+            return -1;
+        }
+        if (a[key] > b[key]) {
+            return 1;
+        }
+        return 0;
+    });
+}
+
+function uniqBy(items, key) {
+    const set = new Set();
+    return items.filter((item) => {
+        if (set.has(item[key])) {
+            return false;
+        }
+        set.add(item[key]);
+        return true;
+    });
+}
+
 module.exports = {
     post,
     getAccessToken,
     uuid,
+    sortBy,
+    uniqBy,
 };
